Move calendar importing logic to DataModel.kt
Makes more sense to have it there, next to the calendar exporting logic.
