+++ app/src/main/java/hu/vmiklos/plees_tracker/DataModel.kt
@@ -19,6 +19,7 @@ import androidx.room.Room
 import androidx.room.migration.Migration
 import androidx.sqlite.db.SupportSQLiteDatabase
 import hu.vmiklos.plees_tracker.calendar.CalendarExport
+import hu.vmiklos.plees_tracker.calendar.CalendarImport
 import java.io.BufferedReader
 import java.io.IOException
 import java.io.InputStreamReader
@@ -168,6 +169,32 @@ object DataModel {
         }
 
         val text = context.getString(R.string.import_success)
+        val duration = Toast.LENGTH_SHORT
+        val toast = Toast.makeText(context, text, duration)
+        toast.show()
+    }
+
+    suspend fun importDataFromCalendar(context: Context, calendarId: String) {
+        // Get all sleeps after the last import, or all of them if there was no previous import.
+        val now = Calendar.getInstance().time
+        var lastCalendarImport = preferences.getLong("last_calendar_import", 0)
+
+        // Query the calendar for events
+        var sleepList = CalendarImport.queryForEvents(
+            context, calendarId
+        ).map(CalendarImport::mapEventToSleep)
+        sleepList = sleepList.filter { it.stop > lastCalendarImport }
+
+        // Insert the list of Sleep into DB
+        insertSleeps(sleepList)
+
+        // Remember the current time for the last import.
+        val editor = preferences.edit()
+        editor.putLong("last_calendar_import", now.time)
+        editor.apply()
+
+        // Show how many sleeps were imported.
+        val text = String.format(context.getString(R.string.imported_items), sleepList.size)
         val duration = Toast.LENGTH_SHORT
         val toast = Toast.makeText(context, text, duration)
         toast.show()
